Trait sui_json_rpc_api::BridgeReadApiClient

pub trait BridgeReadApiClient: ClientT {
    // Provided methods
    fn get_latest_bridge<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<BridgeSummary, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_bridge_object_initial_shared_version<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
}

Client implementation for the BridgeReadApi RPC API.

Provided Methods

fn get_latest_bridge<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<BridgeSummary, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the latest BridgeSummary

fn get_bridge_object_initial_shared_version<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<u64, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the initial shared version of the bridge object, usually for the purpose of constructing an ObjectArg in a transaction.

Object Safety

This trait is not object safe.