Trait sui_json_rpc_api::ExtendedApiClient

pub trait ExtendedApiClient: ClientT {
    // Provided methods
    fn get_epochs<'life0, 'async_trait>(
        &'life0 self,
        cursor: Option<BigInt<u64>>,
        limit: Option<usize>,
        descending_order: Option<bool>
    ) -> Pin<Box<dyn Future<Output = RpcResult<EpochPage>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_epoch_metrics<'life0, 'async_trait>(
        &'life0 self,
        cursor: Option<BigInt<u64>>,
        limit: Option<usize>,
        descending_order: Option<bool>
    ) -> Pin<Box<dyn Future<Output = RpcResult<EpochMetricsPage>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_current_epoch<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<EpochInfo>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn query_objects<'life0, 'async_trait>(
        &'life0 self,
        query: SuiObjectResponseQuery,
        cursor: Option<CheckpointedObjectID>,
        limit: Option<usize>
    ) -> Pin<Box<dyn Future<Output = RpcResult<QueryObjectsPage>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_network_metrics<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<NetworkMetrics>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_move_call_metrics<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<MoveCallMetrics>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_latest_address_metrics<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<AddressMetrics>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_checkpoint_address_metrics<'life0, 'async_trait>(
        &'life0 self,
        checkpoint: u64
    ) -> Pin<Box<dyn Future<Output = RpcResult<AddressMetrics>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_all_epoch_address_metrics<'life0, 'async_trait>(
        &'life0 self,
        descending_order: Option<bool>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<AddressMetrics>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_total_transactions<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<BigInt<u64>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
}

Client implementation for the ExtendedApi RPC API.

Provided Methods

fn get_epochs<'life0, 'async_trait>( &'life0 self, cursor: Option<BigInt<u64>>, limit: Option<usize>, descending_order: Option<bool> ) -> Pin<Box<dyn Future<Output = RpcResult<EpochPage>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return a list of epoch info

fn get_epoch_metrics<'life0, 'async_trait>( &'life0 self, cursor: Option<BigInt<u64>>, limit: Option<usize>, descending_order: Option<bool> ) -> Pin<Box<dyn Future<Output = RpcResult<EpochMetricsPage>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return a list of epoch metrics, which is a subset of epoch info

fn get_current_epoch<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<EpochInfo>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return current epoch info

fn query_objects<'life0, 'async_trait>( &'life0 self, query: SuiObjectResponseQuery, cursor: Option<CheckpointedObjectID>, limit: Option<usize> ) -> Pin<Box<dyn Future<Output = RpcResult<QueryObjectsPage>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return the list of queried objects. Note that this is an enhanced full node only api.

fn get_network_metrics<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<NetworkMetrics>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return Network metrics

fn get_move_call_metrics<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<MoveCallMetrics>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return Network metrics

fn get_latest_address_metrics<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<AddressMetrics>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Address related metrics

fn get_checkpoint_address_metrics<'life0, 'async_trait>( &'life0 self, checkpoint: u64 ) -> Pin<Box<dyn Future<Output = RpcResult<AddressMetrics>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

fn get_all_epoch_address_metrics<'life0, 'async_trait>( &'life0 self, descending_order: Option<bool> ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<AddressMetrics>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

fn get_total_transactions<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<BigInt<u64>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,