sui_json_rpc_api
Trait MoveUtilsServer

pub trait MoveUtilsServer:
    Sized
    + Send
    + Sync
    + 'static {
    // Required methods
    fn get_move_function_arg_types<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module: String,
        function: String,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<MoveFunctionArgType>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_normalized_move_modules_by_package<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<String, SuiMoveNormalizedModule>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_normalized_move_module<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module_name: String,
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedModule>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_normalized_move_struct<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module_name: String,
        struct_name: String,
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedStruct>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_normalized_move_function<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module_name: String,
        function_name: String,
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedFunction>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the MoveUtils RPC API.

Required Methods

fn get_move_function_arg_types<'life0, 'async_trait>( &'life0 self, package: ObjectID, module: String, function: String, ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<MoveFunctionArgType>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the argument types of a Move function, based on normalized Type.

fn get_normalized_move_modules_by_package<'life0, 'async_trait>( &'life0 self, package: ObjectID, ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<String, SuiMoveNormalizedModule>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return structured representations of all modules in the given package

fn get_normalized_move_module<'life0, 'async_trait>( &'life0 self, package: ObjectID, module_name: String, ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedModule>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return a structured representation of Move module

fn get_normalized_move_struct<'life0, 'async_trait>( &'life0 self, package: ObjectID, module_name: String, struct_name: String, ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedStruct>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return a structured representation of Move struct

fn get_normalized_move_function<'life0, 'async_trait>( &'life0 self, package: ObjectID, module_name: String, function_name: String, ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedFunction>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return a structured representation of Move function

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.